fixed update problem
fixed update problem. the problem was runners was filled with the old runner objects. update it would commit, but not update the runner object in the list. added get_athlete to update the runner and replaced it in the runners list.

an alternate solution is to add get_athlete inside of update_athlete.

may change in the future